requirements.txt
asgiref==3.2.5
astroid
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-storages==1.9.1
django-upload-validator==1.1.5
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
wrapt==1.12.1
django-heroku==0.3.1
social-auth-app-django==3.4.0
social-auth-core==3.3.3
